--- client/src/Components/DropDown.css
@@ -1,7 +1,0 @@
-.section {
-
-    justify-content: center;
-    display: block;
-    font-size: large;
-
-}
